IGNITE-29044 Move BinaryWriterSchemaHolder to binary-impl
